refactor(execution): rename symbolTable folder to scope
