+++ src/@types/scope.d.ts
