+++ src/execution/scope/utils.spec.ts
